Removed the 'iemsi user/password' option.
It was always the wrong way around - the username/pw for the connection should've been specified and
not the IEMSI one. That was the "real" user name - the other was just for the connection.

Now when connecting to a ssh bbs that requires a differnt username/password the url sheme should be used:

user:password@URL

The user:password is taken for connection in preference and user/password specified in an address will be used for everything else.